Changed benchmarking score calculation

+++ src/main/java/com/orbix/bench/IBenchmark.java
@@ -8,5 +8,5 @@ public interface IBenchmark
      */
     void initialize(Object... params) throws Exception;
     void run() throws Exception;
-    double getResult();
+    long getResult();
 }
